% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticateConnection.R
\name{authenticateConnection}
\alias{authenticateConnection}
\title{Authenticate connection}
\usage{

  authenticateConnection(url = "https://webapi.asianodds88.com/AsianOddsService/Login?",
  username, password)
}
\arguments{
\item{password}{}
}
\description{
Authenticate connection
}
