% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeeds.R
\name{getFeeds}
\alias{getFeeds}
\title{Retrieves the feed or odds from the system.}
\usage{
getFeeds(url, token, leagueId, sportsType = 1, marketTypeId = 2,
  oddsFormat = "00", since = 1473254981365)
}
\arguments{
\item{since}{}
}
\description{
Retrieves the feed or odds from the system.
}
\references{
\url{https://www.asianodds88.com/documentation.aspx/#4-4-getfeeds}
}
