% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHistoryStatement.R
\name{getHistoryStatement}
\alias{getHistoryStatement}
\title{Retrieves the statement history or odds from the system}
\usage{
getHistoryStatement(url, token, from = "01/01/2016", to = "08/06/2018",
  bookies = "IBC,SBO,SIN,PIN,ISN,GA",
  shouldHideTransactionData = "false")
}
\arguments{
\item{url, token}{see [httr::GET()].}

\item{from, to}{dDate string, mm/dd/yyy.}

\item{bookies}{comma seperated string of bookie ids.}

\item{shouldHideTransactionData}{string with a logical.}
}
\description{
Retrieves the statement history or odds from the system
}
\references{
<https://www.asianodds88.com/documentation.aspx/#3-2-gethistorystatement>
}
