% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLeagueOdds.R
\name{getLeagueOdds}
\alias{getLeagueOdds}
\title{Title}
\usage{
getLeagueOdds(leagueGames, oddsMarket = c("FullTimeOneXTwo", "FullTimeHdp",
  "FullTimeOu"))
}
\arguments{
\item{leagueGames}{}
}
\description{
Title
}
